# docs/openapi/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Nutrition Assistant API",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/v1/settings": {
      "get": {
        "tags": [
          "settings"
        ],
        "summary": "Get Settings",
        "operationId": "get_settings_api_v1_settings_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "type": "object",
                  "title": "Response Get Settings Api V1 Settings Get"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      },
      "patch": {
        "tags": [
          "settings"
        ],
        "summary": "Update Settings",
        "operationId": "update_settings_api_v1_settings_patch",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UserSettingsUpdateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "type": "object",
                  "title": "Response Update Settings Api V1 Settings Patch"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
...
    "/api/v1/uploads/init": {
      "post": {
        "tags": [
          "uploads"
        ],
        "summary": "Init Upload",
        "operationId": "init_upload_api_v1_uploads_init_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UploadInitRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "type": "object",
                  "title": "Response Init Upload Api V1 Uploads Init Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/v1/uploads/complete": {
      "post": {
        "tags": [
          "uploads"
        ],
        "summary": "Complete Upload",
        "operationId": "complete_upload_api_v1_uploads_complete_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UploadCompleteRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "type": "object",
                  "title": "Response Complete Upload Api V1 Uploads Complete Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
...
    "/api/v1/recommendations": {
      "get": {
        "tags": [
          "recommendations"
        ],
        "summary": "List Recommendations",
        "operationId": "list_recommendations_api_v1_recommendations_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "default": 1,
              "title": "Page"
            }
          },
          {
            "name": "page_size",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "default": 20,
              "title": "Page Size"
            }
          },
          {
            "name": "status",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/RecommendationStatus"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Status"
            }
          },
          {
            "name": "recommendation_type",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/RecommendationType"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Recommendation Type"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RecommendationListEnvelopeResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/recommendations/{recommendation_id}/status": {
      "patch": {
        "tags": [
          "recommendations"
        ],
        "summary": "Set Status",
        "operationId": "set_status_api_v1_recommendations__recommendation_id__status_patch",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "recommendation_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Recommendation Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RecommendationUpdateStatusRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Set Status Api V1 Recommendations  Recommendation Id  Status Patch"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "LanguageCode": {
        "type": "string",
        "enum": [
          "en",
          "es",
          "de",
          "fr",
          "ru"
        ],
        "title": "LanguageCode"
      },
...
      "PaginationMeta": {
        "properties": {
          "page": {
            "type": "integer",
            "minimum": 1.0,
            "title": "Page"
          },
          "page_size": {
            "type": "integer",
            "maximum": 200.0,
            "minimum": 1.0,
            "title": "Page Size"
          },
          "total": {
            "type": "integer",
            "minimum": 0.0,
            "title": "Total"
          },
          "total_pages": {
            "type": "integer",
            "minimum": 0.0,
            "title": "Total Pages"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "page",
          "page_size",
          "total",
          "total_pages"
        ],
        "title": "PaginationMeta"
      },
      "RecommendationListDataResponse": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/RecommendationListItemResponse"
            },
            "type": "array",
            "title": "Items"
          },
          "meta": {
            "$ref": "#/components/schemas/PaginationMeta"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "items",
          "meta"
        ],
        "title": "RecommendationListDataResponse"
      },
      "RecommendationListEnvelopeResponse": {
        "properties": {
          "ok": {
            "type": "boolean",
            "const": true,
            "title": "Ok",
            "default": true
          },
          "message_key": {
            "type": "string",
            "title": "Message Key"
          },
          "data": {
            "$ref": "#/components/schemas/RecommendationListDataResponse"
          },
          "error": {
            "type": "null",
            "title": "Error"
          },
          "meta": {
            "additionalProperties": true,
            "type": "object",
            "title": "Meta"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "message_key",
          "data"
        ],
        "title": "RecommendationListEnvelopeResponse"
      },
      "RecommendationListItemResponse": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "type": {
            "type": "string",
            "title": "Type"
          },
          "title": {
            "type": "string",
            "title": "Title"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "id",
          "status",
          "type",
          "title"
        ],
        "title": "RecommendationListItemResponse"
      },
      "RecommendationStatus": {
        "type": "string",
        "enum": [
          "pending",
          "ready",
          "dismissed",
          "applied"
        ],
        "title": "RecommendationStatus"
      },
      "RecommendationType": {
        "type": "string",
        "enum": [
          "meal_adjustment",
          "daily_summary",
          "goal_alignment",
          "hydration"
        ],
        "title": "RecommendationType"
      },
      "RecommendationUpdateStatusRequest": {
        "properties": {
          "status": {
            "$ref": "#/components/schemas/RecommendationStatus"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "status"
        ],
        "title": "RecommendationUpdateStatusRequest"
      },
...
      "UnitSystem": {
        "type": "string",
        "enum": [
          "metric",
          "imperial"
        ],
        "title": "UnitSystem"
      },
      "UploadCompleteRequest": {
        "properties": {
          "upload_id": {
            "type": "string",
            "title": "Upload Id"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "upload_id"
        ],
        "title": "UploadCompleteRequest"
      },
      "UploadInitRequest": {
        "properties": {
          "filename": {
            "type": "string",
            "maxLength": 255,
            "minLength": 1,
            "title": "Filename"
          },
          "mime_type": {
            "type": "string",
            "maxLength": 128,
            "minLength": 3,
            "title": "Mime Type"
          },
          "file_size": {
            "type": "integer",
            "exclusiveMinimum": 0.0,
            "title": "File Size"
          },
          "sha256": {
            "type": "string",
            "maxLength": 64,
            "minLength": 64,
            "title": "Sha256"
          },
          "meal_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Meal Id"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "filename",
          "mime_type",
          "file_size",
          "sha256"
        ],
        "title": "UploadInitRequest"
      },
...
      "UserSettingsUpdateRequest": {
        "properties": {
          "language": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/LanguageCode"
              },
              {
                "type": "null"
              }
            ]
          },
          "unit_system": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/UnitSystem"
              },
              {
                "type": "null"
              }
            ]
          },
          "notifications_enabled": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Notifications Enabled"
          },
          "meal_reminder_enabled": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Meal Reminder Enabled"
          },
          "privacy_flags": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Privacy Flags"
          },
          "feature_flags": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Feature Flags"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "title": "UserSettingsUpdateRequest"
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}